+++ api/src/main/java/org/pmiops/workbench/cdr/ConceptBigQueryService.java
@@ -10,8 +10,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import org.pmiops.workbench.api.BigQueryService;
-import org.pmiops.workbench.cdr.dao.ConceptService;
-import org.pmiops.workbench.cdr.dao.ConceptService.ConceptIds;
+import org.pmiops.workbench.concept.ConceptService;
+import org.pmiops.workbench.concept.ConceptService.ConceptIds;
 import org.pmiops.workbench.config.CdrBigQuerySchemaConfigService;
 import org.pmiops.workbench.config.CdrBigQuerySchemaConfigService.ConceptColumns;
 import org.pmiops.workbench.model.SurveyAnswerResponse;
